break away API functions for building tests
